+++ PaymentService.cs
@@ -4,7 +4,7 @@ namespace Putn
 {
     public class PaymentService : IPaymentService
     {
-        public void Charge(int memberID, decimal totalPayable)
+        public void Charge(int memberID, double totalPayable)
         {
             throw new InvalidOperationException($"{nameof(PaymentService)} I.O. is nice but not for tests");
         }
